fix , add train/test divide
